+++ igdiscover/utils.py
@@ -207,8 +207,10 @@ class Config:
 		self.rename = True
 		self.seed = 1
 		self.exact_copies = None
-		self.pre_germline_filter = dict(unique_cdr3s=2, unique_js=2, check_motifs=False, whitelist=True, cluster_size=0, differences=2)
-		self.germline_filter = dict(unique_cdr3s=5, unique_js=3, check_motifs=False, whitelist=True, cluster_size=100, differences=2)
+		self.pre_germline_filter = dict(unique_cdr3s=2, unique_js=2, check_motifs=False,
+			whitelist=True, cluster_size=0, differences=2, allow_stop=True)
+		self.germline_filter = dict(unique_cdr3s=5, unique_js=3, check_motifs=False,
+			whitelist=True, cluster_size=100, differences=2, allow_stop=False)
 		self.library_name = os.path.basename(os.getcwd())
 
 		self.read_from(file)